refactor(tool-proxy): extract the URL allowlist matcher into its own module
The line ratchet failed: main.rs went to 5031 against a 4975 ceiling, 56 lines
over, because the previous commit added the control-character guard and its
tests. The gate says what to do about it -- "Extract code into sub-modules to
stay under the limit" -- and this is the same move that produced `mediation.rs`.

`url_glob_match` is a coherent unit on its own: the one place that decides
whether an outbound URL is on the allowlist, and worth being able to read
without the rest of the proxy around it. main.rs drops to 4913.

Also fixes the caller in `web_fetch_policy.rs`. Worth recording how it was
found: my grep for callers used a shell glob that expanded wrong and silently
matched nothing, so I concluded there were none. The compiler disagreed. The
lesson is not "grep harder" -- it is that a search returning zero hits and a
search that failed to run look identical, and only one of them is evidence.

No behaviour change; the guard and its three tests move unmodified. Ratchet
passes at 4913/4975, 373 tests green, clippy clean under both `-p` and
`--all-targets`.
